% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-rF2C.Rmd.
\name{F2C}
\alias{F2C}
\title{Convert a Fahrenheit temperature to Celsius}
\usage{
F2C(temp)
}
\arguments{
\item{temp}{A temperature measurement on the Fahrenheit temperature scale}
}
\description{
Convert a Fahrenheit temperature to Celsius
}