% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-rF2C.Rmd.
\docType{package}
\name{rF2C-package}
\alias{rF2C}
\alias{rF2C-package}
\title{My great package}
\description{
Here's an overview of the package.
}
\seealso{
\code{\link{F2C}}
}
\author{
\strong{Maintainer}: First Last \email{you@gmail.com}

}
\keyword{internal}